+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/Evaluator.java
@@ -34,9 +34,8 @@ import com.github.javaparser.ast.type.UnknownType;
 import com.github.javaparser.ast.visitor.VoidVisitorAdapter;
 import com.github.javaparser.resolution.UnsolvedSymbolException;
 
-import sa.com.cloudsolutions.antikythera.evaluator.functional.ConsumerEvaluator;
 import sa.com.cloudsolutions.antikythera.evaluator.functional.FunctionEvaluator;
-import sa.com.cloudsolutions.antikythera.evaluator.functional.FunctionalEvaluator;
+import sa.com.cloudsolutions.antikythera.evaluator.functional.FPEvaluator;
 import sa.com.cloudsolutions.antikythera.exception.AUTException;
 import sa.com.cloudsolutions.antikythera.exception.AntikytheraException;
 import sa.com.cloudsolutions.antikythera.parser.AbstractCompiler;
@@ -70,7 +69,6 @@ import java.io.File;
 import java.util.Optional;
 import java.util.Set;
 import java.util.Stack;
-import java.util.function.Function;
 
 /**
  * Expression evaluator engine.
@@ -230,7 +228,7 @@ public class Evaluator {
     }
 
     private Variable createLambdaExpression(LambdaExpr lambdaExpr) throws ReflectiveOperationException {
-        FunctionalEvaluator eval = FunctionEvaluator.create(lambdaExpr, this);
+        FPEvaluator eval = FunctionEvaluator.create(lambdaExpr, this);
 
         Variable v = new Variable(eval);
         v.setType(new UnknownType());
