+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/functional/FPEvaluator.java
@@ -1,0 +1,138 @@
+package sa.com.cloudsolutions.antikythera.evaluator.functional;
+
+import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.body.MethodDeclaration;
+import com.github.javaparser.ast.expr.Expression;
+import com.github.javaparser.ast.expr.LambdaExpr;
+import com.github.javaparser.ast.stmt.BlockStmt;
+import com.github.javaparser.ast.stmt.ReturnStmt;
+import com.github.javaparser.ast.stmt.Statement;
+import com.github.javaparser.ast.type.UnknownType;
+import sa.com.cloudsolutions.antikythera.evaluator.Evaluator;
+import sa.com.cloudsolutions.antikythera.evaluator.Variable;
+import sa.com.cloudsolutions.antikythera.parser.MCEWrapper;
+
+import java.lang.reflect.Method;
+import java.util.LinkedList;
+import java.util.Optional;
+
+public class FPEvaluator<T> extends Evaluator {
+    protected MethodDeclaration methodDeclaration;
+
+    public FPEvaluator(String className) {
+        super(className);
+    }
+
+    public void setMethod(MethodDeclaration methodDeclaration) {
+        this.methodDeclaration = methodDeclaration;
+    }
+
+    @Override
+    public Variable executeMethod(MCEWrapper wrapper) throws ReflectiveOperationException {
+        returnFrom = null;
+        if (wrapper.getMethodName().equals("apply")) {
+            return executeMethod(methodDeclaration);
+        }
+        throw new UnsupportedOperationException("Not yet implemented");
+    }
+
+
+    public static FPEvaluator create(LambdaExpr lambdaExpr, Evaluator enclosure) throws ReflectiveOperationException {
+        // Create a synthetic method from the lambda
+        MethodDeclaration md = new MethodDeclaration();
+
+        BlockStmt body;
+        if (lambdaExpr.getBody().isBlockStmt()) {
+            body = lambdaExpr.getBody().asBlockStmt();
+            md.setBody(body);
+        } else {
+            body = new BlockStmt();
+            body.addStatement(lambdaExpr.getBody());
+            md.setBody(body);
+        }
+        md.setType(new UnknownType());
+        lambdaExpr.getParameters().forEach(md::addParameter);
+
+        if (checkReturnType(enclosure, body, md) ) {
+            FunctionEvaluator eval = new FunctionEvaluator("functional");
+            eval.setMethod(md);
+            return eval;
+        }
+        else {
+            ConsumerEvaluator eval = new ConsumerEvaluator("functional");
+            eval.setMethod(md);
+            return eval;
+        }
+    }
+
+    private static boolean checkReturnType(Evaluator enclosure, BlockStmt body, MethodDeclaration md) throws ReflectiveOperationException {
+        if (!body.findFirst(ReturnStmt.class).isPresent()) {
+            Statement last = body.getStatements().get(body.getStatements().size() - 1);
+            if (last.isExpressionStmt()) {
+                Expression expr = last.asExpressionStmt().getExpression();
+                if (expr.isMethodCallExpr()) {
+                    LinkedList<Expression> chain = Evaluator.findScopeChain(expr.asMethodCallExpr());
+                    if (chain.isEmpty()) {
+                        return checkunscopedMethod(enclosure, body, md, expr, last);
+                    }
+                    else {
+                        return checkScopedMethod(enclosure, body, md, chain, last);
+                    }
+                }
+                else {
+                    addReturnStatement(body, last);
+                    return true;
+                }
+            }
+        }
+        return true;
+    }
+
+    private static boolean checkunscopedMethod(Evaluator enclosure, BlockStmt body, MethodDeclaration md, Expression expr, Statement last) {
+        /*
+         * We are only concerned about finding the return type here so we don't
+         * need to bother with overloading. All overloaded methods are required to
+         * have the same return type.
+         */
+        CompilationUnit cu = enclosure.getCompilationUnit();
+        if (cu != null) {
+            Optional<MethodDeclaration> foundMethod = cu.findFirst(MethodDeclaration.class,
+                    decl -> decl.getNameAsString().equals(expr.asMethodCallExpr().getNameAsString()));
+
+            if (foundMethod.isPresent()) {
+                addReturnStatement(body, last);
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private static boolean checkScopedMethod(Evaluator enclosure, BlockStmt body, MethodDeclaration md, LinkedList<Expression> chain, Statement last) throws ReflectiveOperationException {
+        Variable v = enclosure.evaluateScopeChain(chain);
+        if (v != null) {
+            if (v.getValue() instanceof Evaluator e) {
+
+            }
+            else {
+                Class<?> clz = v.getClazz();
+                for (Method m : clz.getMethods()) {
+                    if (m.getName().equals(md.getNameAsString())) {
+                        if (!m.getReturnType().equals(Void.TYPE)) {
+                            addReturnStatement(body, last);
+                            return true;
+                        }
+                    }
+                }
+            }
+        }
+        return false;
+    }
+
+    private static void addReturnStatement(BlockStmt body, Statement last) {
+        body.remove(last);
+        ReturnStmt returnStmt = new ReturnStmt();
+        returnStmt.setExpression(last.asExpressionStmt().getExpression());
+        body.addStatement(returnStmt);
+    }
+
+}
